debug: won't writeover existed input-xxx.txt

+++ libmasscan/output_masscan_ip.py
@@ -1,5 +1,6 @@
 # encoding=utf-8
 import time
+import os
 import pymongo
 
 
@@ -13,13 +14,13 @@ def output():
         domain_ip_ports = pymongo.MongoClient('localhost', 27017)['virustotal']['domain-ip-ports']
         ips = domain_ip_ports.distinct('ip_str', {'status': '101'})
         filename = 'libmasscan/data/input-%s.txt' % time.strftime('%Y-%m-%d', time.localtime(time.time()))
-
-        with open(filename, 'w') as f:
-            for ip in ips:
-                f.write('%s\n' % str(ip))
-                # 更新status 201，表示已被提取出来，交给masscan处理
-                domain_ip_ports.update({'ip_str': ip}, {'$set': {'status': '201'}})
-        print '[info]Output input.txt'
+        if not os.path.exists(filename):
+            with open(filename, 'w') as f:
+                for ip in ips:
+                    f.write('%s\n' % str(ip))
+                    # 更新status 201，表示已被提取出来，交给masscan处理
+                    domain_ip_ports.update({'ip_str': ip}, {'$set': {'status': '201'}})
+            print '[info]Output input.txt'
 
     except Exception as msg:
         print '[error]%s' % msg
